+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/TestCOWDataSource.scala
@@ -762,7 +762,6 @@ class TestCOWDataSource extends HoodieClientTestBase {
     }
   }
 
-
   @Test def testSchemaNotEqualData(): Unit = {
     val  opts = commonOpts ++ Map("hoodie.avro.schema.validate" -> "true")
     val schema1 = StructType(StructField("_row_key", StringType, true) :: StructField("name", StringType, true)::
@@ -783,125 +782,5 @@ class TestCOWDataSource extends HoodieClientTestBase {
 
     val resultSchema = new StructType(recordsReadDF.schema.filter(p=> !p.name.startsWith("_hoodie")).toArray)
     assertEquals(resultSchema, schema1)
-  }
-
-
-  @Test def testSimpleAutoInferDefaultPartition(): Unit = {
-    val records = recordsToStrings(dataGen.generateInserts("000", 100)).toList
-    val inputDF = spark.read.json(spark.sparkContext.parallelize(records, 2))
-    // Default partition
-    inputDF.write.format("org.apache.hudi")
-      .options(commonOpts)
-      .mode(SaveMode.Overwrite)
-      .save(basePath)
-
-    val supplyBlobs = spark.read.format("org.apache.hudi")
-      .load(basePath + "/*/*/*/")
-    assertEquals(inputDF.count(), supplyBlobs.count())
-
-    // No need to specify basePath/*/*/*
-    val withoutBlobs = spark.read.format("org.apache.hudi")
-      .load(basePath)
-    assertEquals(inputDF.count(), withoutBlobs.count())
-
-    // Testing with trailing backslash basePath/
-    val trailingSlash = spark.read.format("org.apache.hudi")
-      .load(basePath + "/")
-    assertEquals(inputDF.count(), trailingSlash.count())
-  }
-
-  @Test def testSimpleAutoInferCustomPartition(): Unit = {
-    val records = recordsToStrings(dataGen.generateInserts("000", 100)).toList
-    val inputDF = spark.read.json(spark.sparkContext.parallelize(records, 2))
-    // Partition with org.apache.hudi.keygen.CustomKeyGenerator
-    inputDF.write.format("org.apache.hudi")
-      .options(commonOpts)
-      .option(DataSourceWriteOptions.KEYGENERATOR_CLASS_OPT_KEY.key, "org.apache.hudi.keygen.CustomKeyGenerator")
-      .option(DataSourceWriteOptions.PARTITIONPATH_FIELD_OPT_KEY.key, "rider:SIMPLE,begin_lon:SIMPLE,end_lon:SIMPLE")
-      .mode(SaveMode.Overwrite)
-      .save(basePath)
-
-    val supplyBlobs = spark.read.format("org.apache.hudi")
-      .load(basePath + "/*/*/*")
-    assertEquals(inputDF.count(), supplyBlobs.count())
-
-    // No need to specify basePath/*/*/*
-    val withoutBlobs = spark.read.format("org.apache.hudi")
-      .load(basePath)
-    assertEquals(inputDF.count(), withoutBlobs.count())
-
-    val riderVal = inputDF.first().getAs[String]("rider")
-    val beginLonVal = inputDF.first().getAs[Double]("begin_lon")
-    val endLonVal = inputDF.first().getAs[Double]("end_lon")
-
-    val partitionBlob = spark.read.format("org.apache.hudi")
-      .load(basePath + "/" + riderVal + "/" + beginLonVal + "/*")
-    assertEquals(inputDF.filter(col("begin_lon") === beginLonVal).count(),
-      partitionBlob.count())
-
-    val specificPartition = spark.read.format("org.apache.hudi")
-      .load(basePath + "/" + riderVal + "/" + beginLonVal + "/" + endLonVal + "/")
-    assertEquals(inputDF.filter(
-      col("rider") === riderVal
-        && col("begin_lon") === beginLonVal
-        && col("end_lon") === endLonVal).count(),
-      specificPartition.count())
-
-    val wildcardBlob = spark.read.format("org.apache.hudi")
-      .load(basePath + "/*/0.4*/*")
-    assertEquals(inputDF.filter(col("begin_lon").startsWith("0.4")).count(),
-      wildcardBlob.count())
-  }
-
-  @Test def testSimpleAutoInferNoPartition(): Unit = {
-    val records = recordsToStrings(dataGen.generateInserts("000", 100)).toList
-    val inputDF = spark.read.json(spark.sparkContext.parallelize(records, 2))
-
-    // No partition
-    inputDF.write.format("org.apache.hudi")
-      .options(commonOpts)
-      .option(DataSourceWriteOptions.KEYGENERATOR_CLASS_OPT_KEY.key, "org.apache.hudi.keygen.NonpartitionedKeyGenerator")
-      .option(DataSourceWriteOptions.PARTITIONPATH_FIELD_OPT_KEY.key, "")
-      .mode(SaveMode.Overwrite)
-      .save(basePath)
-
-    val supplyBlobs = spark.read.format("org.apache.hudi")
-      .load(basePath + "/*")
-    assertEquals(inputDF.count(), supplyBlobs.count())
-
-    // No need to specify basePath/*/*/*
-    val withoutBlobs = spark.read.format("org.apache.hudi")
-      .load(basePath)
-    assertEquals(inputDF.count(), withoutBlobs.count())
-  }
-
-  @Test def testSimpleAutoInferDatePartition(): Unit = {
-    val records = recordsToStrings(dataGen.generateInserts("000", 100)).toList
-    val inputDF = spark.read.json(spark.sparkContext.parallelize(records, 2))
-
-    inputDF.write.format("org.apache.hudi")
-      .options(commonOpts)
-      .option(DataSourceWriteOptions.KEYGENERATOR_CLASS_OPT_KEY.key, "org.apache.hudi.keygen.CustomKeyGenerator")
-      .option(DataSourceWriteOptions.PARTITIONPATH_FIELD_OPT_KEY.key, "current_ts:TIMESTAMP")
-      .option(Config.TIMESTAMP_TYPE_FIELD_PROP, "EPOCHMILLISECONDS")
-      .option(Config.TIMESTAMP_OUTPUT_DATE_FORMAT_PROP, "yyyyMMdd")
-      .mode(SaveMode.Overwrite)
-      .save(basePath)
-
-    val supplyBlobs = spark.read.format("org.apache.hudi")
-      .load(basePath + "/*/*")
-    assertEquals(inputDF.count(), supplyBlobs.count())
-
-    // No need to specify basePath/*/*/*
-    val withoutBlobs = spark.read.format("org.apache.hudi")
-      .load(basePath)
-    assertEquals(inputDF.count(), withoutBlobs.count())
-
-    // Specify basePath/yyyyMMdd/*
-    val date = new DateTime().toString(DateTimeFormat.forPattern("yyyyMMdd"))
-    val specificDatePartition = spark.read.format("org.apache.hudi")
-      .load(basePath + s"/$date")
-    assertEquals(supplyBlobs.filter(col("_hoodie_partition_path") === date).count(),
-      specificDatePartition.count())
   }
 }
